add rest api

+++ app/Http/Controllers/ProductController.php
@@ -16,7 +16,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::latest()->get();
+        $products = Product::get();
 
         // Untuk Web
         // return view('dashboard-admin/index');
@@ -25,17 +25,22 @@ class ProductController extends Controller
         // return response()->json(['data' => $products]);
 
         // mirip dengan Tipe 1 tapi menggunakan resource dan harus di instal dulu packpage nya 
+
+        if ($products->isEmpty()) {
+            return response()->json(['message' => 'No products found.'], 404);
+        }
+
         return ProductResource::collection($products);
     }
 
     public function viewIndex()
     {
-        $products = Product::get();
+        $products = Product::paginate(10);
         return view('dashboard-admin/product/product',['products' => $products]);
     }
     public function viewInsert()
     {
-        $category =  Category::first()->get();
+        $category =  Category::get();
         return view('dashboard-admin/product/insert-product',['categories' => $category]);
     }
     public function insert(Request $request)
@@ -44,17 +49,6 @@ class ProductController extends Controller
         $validatedData = $request->validate([
             'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048'
         ]);
-
-        // $fileName = null;
-        // if ($request->hasFile('image')) {
-        //     // generate random string untuk nama file
-        //     $fileName = $this->generateRandomString();
-        //     // ambil extension file
-        //     $extension = $request->file('image')->extension();
-
-        //     // Simpan file gambar ke storage
-        //     Storage::putFileAs('product-image', $request->file('image'), $fileName . '.' . $extension);
-        // }
 
         $fileName = null;
         if ($request->hasFile('image')) {
@@ -104,32 +98,31 @@ class ProductController extends Controller
         $product->category_id = $request->input('category');
         $product->stock = $request->input('stock');
 
-        $fileName = null;
         if ($request->hasFile('image')) {
-        // generate random string untuk nama file
             $fileName = $this->generateRandomString();
-
-        // ambil extension file
             $extension = $request->file('image')->extension();
-
-        // Path tujuan di folder public
             $path = public_path('product-image');
 
-        // Pastikan folder ada, jika tidak, buat foldernya
             if (!file_exists($path)) {
                 mkdir($path, 0755, true);
             }
 
-        // Simpan file gambar ke folder public/product-image
-            $request->file('image')->move($path, $fileName . '.' . $extension);
+            if ($product->image) {
+                $oldImagePath = $path . '/' . $product->image;
+                if (file_exists($oldImagePath)) {
+                    unlink($oldImagePath);
+                }
+            }
 
-        // Set nama file yang baru ke dalam database
+            $request->file('image')->move($path, $fileName . '.' . $extension);
             $product->image = $fileName . '.' . $extension;
         }
 
         $product->save();
         return redirect('/product-admin');
     }
+
+
     public function deleteProduct($id)
     {
         $product = Product::findOrFail($id);
@@ -160,27 +153,59 @@ class ProductController extends Controller
     {
         $validated = $request->validate([
             'name' => 'required',
+            'price' => 'required',
             'category_id' => 'required',
             'stock' => 'required',
             'file_img' => 'required',
         ]);
 
-        // Jika file gambar ada
+        $fileName = null;
         if ($request->hasFile('file_img')) {
-            // generate random string untuk nama file
+        // generate random string untuk nama file
             $fileName = $this->generateRandomString();
-            // ambil extension file
+        // ambil extension file
             $extension = $request->file('file_img')->extension();
 
-            // Simpan file gambar ke storage
-            Storage::putFileAs('product-image', $request->file('file_img'), $fileName . '.' . $extension);
+        // Path tujuan di folder public
+            $path = public_path('product-image');
 
-            // Tambahkan nama file yang baru ke dalam array validasi
-            $validated['image'] = $fileName . '.' . $extension;
+        // Pastikan folder ada, jika tidak, buat foldernya
+            if (!file_exists($path)) {
+                mkdir($path, 0755, true);
+            }
+
+         // Simpan file gambar ke folder public/product-image
+            $request->file('file_img')->move($path, $fileName . '.' . $extension);
         }
 
 
-        $product = Product::create($validated);
+        // Simpan data produk ke database
+        $product = Product::create([
+            'name' => $request->name,
+            'price' => $request->price,
+            'category_id' => $request->category_id,
+            'image' => $fileName ? $fileName . '.' . $extension : null,
+            'stock' => $request->stock
+        ]);
+
+
+
+        // // Jika file gambar ada
+        // if ($request->hasFile('file_img')) {
+        //     // generate random string untuk nama file
+        //     $fileName = $this->generateRandomString();
+        //     // ambil extension file
+        //     $extension = $request->file('file_img')->extension();
+
+        //     // Simpan file gambar ke storage
+        //     Storage::putFileAs('product-image', $request->file('file_img'), $fileName . '.' . $extension);
+
+        //     // Tambahkan nama file yang baru ke dalam array validasi
+        //     $validated['image'] = $fileName . '.' . $extension;
+        // }
+
+
+        // $product = Product::create($validated);
 
         return new ProductResource($product); 
         // return response()->json('dd');
@@ -201,9 +226,13 @@ class ProductController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Product $product)
+    public function show($id)
     {
-        $product = Product::findOrFail($product->id);
+        $product = Product::find($id);
+
+        if (!$product) {
+            return response()->json(['message' => 'Product not found.'], 404);
+        }
         
         // Tipe 1 
         // return response()->json(['data' => $product]);
@@ -232,11 +261,32 @@ class ProductController extends Controller
     {
         $validated = $request->validate([
             'name' => 'required',
+            'price' => 'required',
             'category_id' => 'required',
             'stock' => 'required',
+            'file_img' => 'nullable',
         ]);
 
         $product = Product::findOrFail($id);
+        if ($request->hasFile('file_img')) {
+            $fileName = $this->generateRandomString();
+            $extension = $request->file('file_img')->extension();
+            $path = public_path('product-image');
+
+            if (!file_exists($path)) {
+                mkdir($path, 0755, true);
+            }
+
+            if ($product->image) {
+                $oldImagePath = $path . '/' . $product->image;
+                if (file_exists($oldImagePath)) {
+                    unlink($oldImagePath);
+                }
+            }
+
+            $request->file('file_img')->move($path, $fileName . '.' . $extension);
+            $product->image = $fileName . '.' . $extension;
+        }
         $product->update($request->all());
 
         return new ProductResource($product);
@@ -247,7 +297,12 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $product = Product::findOrFail($id);
+        $product = Product::find($id);
+
+        if (!$product) {
+            return response()->json(['message' => 'Product not found.'], 404);
+        }
+        
         $product->delete();
 
         return new ProductResource($product);
